Some cleaning is done. Malloc part of window is not at the beginning. Some functions are split in new files like draw and motion
